requirements.txt
droit
python-telegram-bot